pyproject.toml
[project]
name = "resolve-works"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2,<5.3",
    "django-browser-reload>=1.21.0",
    "wagtail>=7.2",
    "wagtail-markdown>=0.13.0",
    "sentence-transformers>=3.3.0",
    "umap-learn>=0.5.7",
    "nltk>=3.9.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"
